Merge the tables that track dereference, store address and use in functions/syscalls of capability vertices, into a single events table.
Add support to stats to detect capabilities originating in the initial arguments stack.
Fix minor bugs in merge logic.
Add a new type of entry to the capability vertex events table to track every time that the vertex is loaded from memory.

The new events table have two representations. A dict-based representation is used during parsing because it offers high
append performance. A pandas dataframe-based representation is used for lookups.